benchmark/requirements.txt
ragas>=0.2.0
langchain-openai>=0.1.0
datasets>=2.18.0
pyyaml>=6.0
